Split and clean code
Main changes:

* Create Network, Matrix, Flow, Wiring, Log, and VectorGuard files
* Drop Types, Network Tools, Util files
* Move outputRatios into Flow
* Move Exists Balancers wiring generation code into Wiring

Other changes:

* rename outputRatios to networkToFlow
* Generate networks from vectors of connection pairs
* Tidy existsBalancer + existsBalancer testcases
* Use Pair for Wiring Configs
