+++ include/log.hpp
@@ -1,0 +1,12 @@
+// Utilities in the system domain (like logging)
+
+#pragma once
+
+#include <string>
+#include <vector>
+
+// Log a string to console
+void log(std::string message);
+
+// Log a vector<double> to console in full precision
+void log(std::vector<double> row);
