+++ src/network.cpp
@@ -1,0 +1,42 @@
+// Tools for the Network/Matrix domain
+
+#include "network.hpp"
+
+#include <algorithm>
+
+#include "vector_guard.hpp"
+
+Network emptyNetwork(int size) {
+  Network nodes;
+  for (int i = 0; i < size; ++i) {
+    Node* node = new Node;
+    nodes.push_back(node);
+  }
+  return nodes;
+}
+
+void link(Network& nodes, int source, int target) {
+  vectorGuard(nodes, source);
+  vectorGuard(nodes, target);
+  Node* target_node = nodes[target];
+  Node* source_node = nodes[source];
+  source_node->outputs.push_back(target_node);
+  target_node->inputs.push_back(source_node);
+}
+
+Network makeNetwork(const int SIZE, const Links LINKS) {
+  Network network = emptyNetwork(SIZE);
+  for (auto L : LINKS) {
+    link(network, L.first, L.second);
+  }
+  return network;
+}
+
+int nodeNum(Network nodes, Node* node) {
+  for (int node_num = 0; node_num < nodes.size(); ++node_num) {
+    if (node == nodes[node_num]) {
+      return node_num;
+    }
+  }
+  throw "Node not found";
+}
