+++ tests/test_existsBalancer.cpp
@@ -1,7 +1,4 @@
 // Test Runner for exists_balancers
-
-#include <iostream>
-
 #include "catch2/catch.hpp"
 #include "exists_balancer.hpp"
